refactor: move test files into test subfolder
