+++ src/view/secondary-bar/custom-accordion-components/test/tag-accordion-item.test.tsx
@@ -14,9 +14,9 @@ import {
 	ZIMBRA_STANDARD_COLORS
 } from '@zextras/carbonio-ui-commons';
 
-import { TagAccordionItem } from './tag-accordion-item';
-import { setupTest, screen } from '../../../__test__/test-setup';
-import { TEST_SELECTORS } from '../../../constants/test-utils';
+import { setupTest, screen } from '../../../../__test__/test-setup';
+import { TEST_SELECTORS } from '../../../../constants/test-utils';
+import { TagAccordionItem } from '../tag-accordion-item';
 
 const generateTag = (model: Partial<Tag> = {}): Tag => ({
 	id: model.id || faker.string.uuid(),
